+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 ## [Unreleased]
 ### Added
 - iRODS setup code, split of from [test-with-baton](https://github.com/wtsi-hgi/test-with-baton) repository.
+
+### Changed
+- iput helper no longer uses the project directory for temp files.